+++ manticore/manticore.py
@@ -16,17 +16,28 @@ from threading import Timer
 from elftools.elf.elffile import ELFFile
 from elftools.elf.sections import SymbolTableSection
 
-from .core.workspace import ManticoreOutput
 from .core.executor import Executor
-from .core.state import State, TerminateState
 from .core.parser import parse
+from .core.state import State, TerminateState
 from .core.smtlib import solver, ConstraintSet
+from .core.workspace import ManticoreOutput
 from .platforms import linux, decree, windows
-from .utils.helpers import issymbolic
+from .utils.helpers import issymbolic, is_binja_disassembler
 from .utils.nointerrupt import WithKeyboardInterruptAs
+from .utils import log
 
 logger = logging.getLogger('MANTICORE')
 
+def makeBinja(program, disasm, argv, env, symbolic_files, concrete_start=''):
+    constraints = ConstraintSet()
+    logger.info('Loading binary ninja IL from %s', program)
+    platform = linux.SLinux(program,
+                            argv=argv,
+                            envp=env,
+                            symbolic_files=symbolic_files,
+                            disasm=disasm)
+    initial_state = State(constraints, platform)
+    return initial_state
 
 def makeDecree(program, concrete_data=''):
     constraints = ConstraintSet()
@@ -40,17 +51,20 @@ def makeDecree(program, concrete_data=''):
     platform.input.transmit(initial_state.symbolicate_buffer('+'*14, label='RECEIVE'))
     return initial_state
 
-def makeLinux(program, argv, env, symbolic_files, concrete_start = ''):
-    logger.info('Loading program %s', program)
+def makeLinux(program, disasm, argv, env, symbolic_files, concrete_start=''):
+    logger.info('Loading program %s (platform: Linux)', program)
 
     constraints = ConstraintSet()
 
-    platform = linux.SLinux(program, argv=argv, envp=env,
-                            symbolic_files=symbolic_files)
+    platform = linux.SLinux(program,
+                            argv=argv,
+                            envp=env,
+                            symbolic_files=symbolic_files,
+                            disasm=disasm)
     initial_state = State(constraints, platform)
 
     if concrete_start != '':
-        logger.info('Starting with concrete input: {}'.format(concrete_start))
+        logger.info('Starting with concrete input: %s', concrete_start)
 
     for i, arg in enumerate(argv):
         argv[i] = initial_state.symbolicate_buffer(arg, label='ARGV%d' % (i+1))
@@ -66,15 +80,16 @@ def makeLinux(program, argv, env, symbolic_files, concrete_start = ''):
     platform.input.write(concrete_start)
 
     #set stdin input...
-    platform.input.write(initial_state.symbolicate_buffer('+'*256, label='STDIN'))
+    platform.input.write(initial_state.symbolicate_buffer('+' * 256,
+                                                          label='STDIN'))
 
-    return initial_state 
+    return initial_state
 
 
 def makeWindows(args):
     assert args.size is not None, "Need to specify buffer size"
     assert args.buffer is not None, "Need to specify buffer base address"
-    logger.debug('Loading program %s', args.programs)
+    logger.debug('Loading program %s (platform: Windows)', args.programs)
     additional_context = None
     if args.context:
         with open(args.context, "r") as addl_context_file:
@@ -92,7 +107,7 @@ def makeWindows(args):
     buf_str = "".join(platform.current.read_bytes(data_ptr, data_size))
     logger.debug('Original buffer: %s', buf_str.encode('hex'))
 
-    offset = args.offset 
+    offset = args.offset
     concrete_data = args.data.decode('hex')
     assert data_size >= offset + len(concrete_data)
     size = min(args.maxsymb, data_size - offset - len(concrete_data))
@@ -131,6 +146,7 @@ def binary_type(path):
     else:
         raise NotImplementedError("Binary {} not supported. Magic bytes: 0x{}".format(path, binascii.hexlify(magic)))
 
+
 class Manticore(object):
     '''
     The central analysis object.
@@ -140,15 +156,16 @@ class Manticore(object):
     :type args: list[str]
     :ivar dict context: Global context for arbitrary data storage
     '''
-
-
-    def __init__(self, binary_path, args=None):
+    def __init__(self, binary_path, args=None, disasm='capstone'):
         assert os.path.isfile(binary_path)
 
         args = [] if args is None else args
 
+        self._check_disassembler_present(disasm)
+        self._disasm = disasm
         self._binary = binary_path
         self._binary_type = binary_type(binary_path)
+        # yan's comment
         self._argv = args # args.programs[1:]
         self._env = {}
         # Will be set to a temporary directory if not set before running start()
@@ -170,7 +187,6 @@ class Manticore(object):
         self._dumpafter = 0
         self._maxstates = 0
         self._maxstorage = 0
-        self._verbosity = 0
         self._symbolic_files = [] # list of string
         self._executor = None
         #Executor wide shared context
@@ -181,7 +197,7 @@ class Manticore(object):
         if self._binary_type == 'ELF':
             self._binary_obj = ELFFile(file(self._binary))
 
-        self._init_logging()
+        log.init_logging()
 
     @property
     def context(self):
@@ -191,18 +207,33 @@ class Manticore(object):
         else:
             logger.warning("Using shared context without a lock")
             return self._executor._shared_context
-        
+
 
     @contextmanager
-    def locked_context(self, key=None, default=list):
-        ''' It refers to the manticore shared context 
-            It needs a lock. Its used like this:
+    def locked_context(self, key=None, value_type=list):
+        """
+        A context manager that provides safe parallel access to the global Manticore context.
+        This should be used to access the global Manticore context
+        when parallel analysis is activated. Code within the `with` block is executed
+        atomically, so access of shared variables should occur within.
 
-            with m.context() as context:
-                vsited = context['visited']
+        Example use::
+
+            with m.locked_context() as context:
+                visited = context['visited']
                 visited.append(state.cpu.PC)
                 context['visited'] = visited
-        '''
+
+        Optionally, parameters can specify a key and type for the object paired to this key.::
+
+            with m.locked_context('feature_list', list) as feature_list:
+                feature_list.append(1)
+
+        :param object key: Storage key
+        :param value_type: type of value associated with key
+        :type value_type: list or dict or set
+        """
+
         @contextmanager
         def _real_context():
             if self._executor is None:
@@ -215,36 +246,10 @@ class Manticore(object):
             if key is None:
                 yield context
             else:
-                assert default in (list, dict, set)
-                ctx = context.get(key, None)
-                if ctx is None:
-                    ctx = default()
+                assert value_type in (list, dict, set)
+                ctx = context.get(key, value_type())
                 yield ctx
                 context[key] = ctx
-
-    def _init_logging(self): 
-
-        def loggerSetState(logger, stateid):
-            logger.filters[0].stateid = stateid
-
-        class ContextFilter(logging.Filter):
-            '''
-            This is a filter which injects contextual information into the log.
-            '''
-            def filter(self, record):
-                if hasattr(self, 'stateid') and isinstance(self.stateid, int):
-                    record.stateid = '[%d]' % self.stateid
-                else:
-                    record.stateid = ''
-                return True
-
-        ctxfilter = ContextFilter()
-
-        logging.basicConfig(format='%(asctime)s: [%(process)d]%(stateid)s %(name)s:%(levelname)s: %(message)s', stream=sys.stdout, level=logging.ERROR)
-
-        for loggername in ['MANTICORE', 'VISITOR', 'EXECUTOR', 'CPU', 'REGISTERS', 'SMT', 'MEMORY', 'PLATFORM']:
-            logging.getLogger(loggername).addFilter(ctxfilter)
-            logging.getLogger(loggername).setState = types.MethodType(loggerSetState, logging.getLogger(loggername))
 
     # XXX(yan): args is a temporary hack to include while we continue moving
     # non-Linux platforms to new-style arg handling.
@@ -292,40 +297,23 @@ class Manticore(object):
     def maxstorage(self):
         return self._maxstorage
 
-    @maxstorage.setter
-    def maxstorage(self, max_storage):
-        self._maxstorage = max_storage
-
     @property
     def verbosity(self):
-        '''
-        Convenience interface for setting logging verbosity to one of several predefined
-        logging presets. Valid values: 0-5
-        '''
-        return self._verbosity
+        """Convenience interface for setting logging verbosity to one of
+        several predefined logging presets. Valid values: 0-5.
+        """
+        return log.manticore_verbosity
 
     @verbosity.setter
     def verbosity(self, setting):
-        zero = map(lambda x: (x, logging.ERROR),
-                   ['MANTICORE', 'VISITOR', 'EXECUTOR', 'CPU', 'REGISTERS', 'SMT', 'MEMORY', 'PLATFORM'])
-        levels = [zero,
-                  [('MANTICORE', logging.INFO), ('EXECUTOR', logging.INFO)],
-                  [('PLATFORM', logging.DEBUG)],
-                  [('MEMORY', logging.DEBUG), ('CPU', logging.DEBUG)],
-                  [('REGISTERS', logging.DEBUG)],
-                  [('SMT', logging.DEBUG)]]
+        """A call used to modify the level of output verbosity
+        :param int setting: the level of verbosity to be used
+        """
+        log.set_verbosity(setting)
 
-        # Takes a value and ensures it's in a certain range
-        def clamp(val, minimum, maximum):
-            return sorted((minimum, val, maximum))[1]
-
-        clamped = clamp(setting, 0, len(levels) - 1)
-        if clamped != setting:
-            logger.debug("%s not between 0 and %d, forcing to %d", setting, len(levels) - 1, clamped)
-        for level in range(clamped + 1):
-            for log_type, log_level in levels[level]:
-                logging.getLogger(log_type).setLevel(log_level)
-        self._verbosity = clamped
+    @maxstorage.setter
+    def maxstorage(self, max_storage):
+        self._maxstorage = max_storage
 
     def hook(self, pc):
         '''
@@ -383,10 +371,16 @@ class Manticore(object):
             return symbols[0].entry['st_value']
 
     def _make_state(self, path):
-        if self._binary_type == 'ELF':
+        if self._binary_type == 'BinaryNinjaIL' or self._disasm == "binja-il":
+            # Binary Ninja
+            env = ['%s=%s' % (k, v) for k, v in self._env.items()]
+            state = makeBinja(self._binary, self._disasm, self._argv, env,
+                              self._symbolic_files, self._concrete_data)
+        elif self._binary_type == 'ELF':
             # Linux
-            env = ['%s=%s'%(k,v) for k,v in self._env.items()]
-            state = makeLinux(self._binary, self._argv, env, self._symbolic_files, self._concrete_data)
+            env = ['%s=%s' % (k, v) for k, v in self._env.items()]
+            state = makeLinux(self._binary, self._disasm, self._argv, env,
+                              self._symbolic_files, self._concrete_data)
         elif self._binary_type == 'PE':
             # Windows
             state = makeWindows(self._args)
@@ -397,7 +391,6 @@ class Manticore(object):
             raise NotImplementedError("Binary {} not supported.".format(path))
 
         return state
-        
 
     @property
     def policy(self):
@@ -463,7 +456,7 @@ class Manticore(object):
         else: raise "Unsupported architecture: %s"%(arch, )
 
         return self._arch
-        
+
 
     def _start_workers(self, num_processes, profiling=False):
         assert num_processes > 0, "Must have more than 0 worker processes"
@@ -488,13 +481,16 @@ class Manticore(object):
         else:
             target = self._executor.run
 
-        for _ in range(num_processes):
-            p = Process(target=target, args=())
-            self._workers.append(p)
-            p.start()
+        if self._disasm == 'binja-il':
+            target()
+        else:
+            for _ in range(num_processes):
+                p = Process(target=target, args=())
+                self._workers.append(p)
+                p.start()
 
     def _join_workers(self):
-        with WithKeyboardInterruptAs(self._executor.shutdown):    
+        with WithKeyboardInterruptAs(self._executor.shutdown):
             while len(self._workers) > 0:
                 w = self._workers.pop().join()
 
@@ -547,10 +543,10 @@ class Manticore(object):
                 self._assertions[pc] = ' '.join(line.split(' ')[1:])
 
     def _store_state_callback(self, state, state_id):
-        logger.debug("store state %r", state_id)
+        logger.info("store state %r", state_id)
 
     def _load_state_callback(self, state, state_id):
-        logger.debug("load state %r", state_id)
+        logger.info("load state %r", state_id)
 
     def _terminate_state_callback(self, state, state_id, ex):
         #aggregates state statistics into exceutor statistics. FIXME split
@@ -564,7 +560,10 @@ class Manticore(object):
             manticore_context['visited'] = manticore_visited.union(state_visited)
 
             manticore_instructions_count = manticore_context.get('instructions_count', 0)
-            manticore_context['instructions_count'] = manticore_instructions_count + state_instructions_count 
+            manticore_context['instructions_count'] = manticore_instructions_count + state_instructions_count
+
+    def _forking_state_callback(self, state, expression, value, policy):
+        state.record_branch(value)
 
     def _fork_state_callback(self, state, expression, values, policy):
         state_visited = state.context.get('visited_since_last_fork', set())
@@ -573,9 +572,7 @@ class Manticore(object):
             manticore_context['visited'] = manticore_visited.union(state_visited)
         state.context['visited_since_last_fork'] = set()
 
-        logger.debug("About to store state %r %r %r", state, expression, values, policy)
-
-    def _read_register_callback(self, state, reg_name, value): 
+    def _read_register_callback(self, state, reg_name, value):
         logger.debug("Read Register %r %r", reg_name, value)
 
     def _write_register_callback(self, state, reg_name, value):
@@ -590,7 +587,6 @@ class Manticore(object):
     def _decode_instruction_callback(self, state, pc):
         logger.debug("Decoding stuff instruction not available")
 
-
     def _emulate_instruction_callback(self, state, instruction):
         logger.debug("About to emulate instruction")
 
@@ -600,20 +596,21 @@ class Manticore(object):
     def _execute_instruction_callback(self, state, instruction):
         address = state.cpu.PC
         if not issymbolic(address):
+            state.context.setdefault('trace', list()).append(address)
             state.context.setdefault('visited_since_last_fork', set()).add(address)
             state.context.setdefault('visited', set()).add(address)
             count = state.context.get('instructions_count', 0)
             state.context['instructions_count'] = count + 1
 
 
-    def _generate_testcase_callback(self, state, message = 'Testcase generated'):
+    def _generate_testcase_callback(self, state, name, message):
         '''
         Create a serialized description of a given state.
         :param state: The state to generate information about
         :param message: Accompanying message
         '''
-        testcase_id = self._output.save_testcase(state, message)
-        logger.debug("Generated testcase No. %d - %s", testcase_id, message)
+        testcase_id = self._output.save_testcase(state, name, message)
+        logger.info("Generated testcase No. {} - {}".format(testcase_id, message))
 
     def _produce_profiling_data(self):
         class PstatsFormatted:
@@ -662,15 +659,11 @@ class Manticore(object):
             for entry in sorted(executor_visited):
                 f.write('0:{:08x}\n'.format(entry))
 
-        #if self.coverage_file is not None:
-        #    import shutil
-        #    shutil.copyfile('visited.txt', self.coverage_file)
-
         with self._output.save_stream('command.sh') as f:
             f.write(' '.join(sys.argv))
 
-        instructions_count = _shared_context.get('instructions_count',0)
-        elapsed = time.time()-self._time_started
+        instructions_count = _shared_context.get('instructions_count', 0)
+        elapsed = time.time() - self._time_started
         logger.info('Results in %s', self._output.uri)
         logger.info('Instructions executed: %d', instructions_count)
         logger.info('Coverage: %d different instructions executed', len(executor_visited))
@@ -689,7 +682,7 @@ class Manticore(object):
         assert not self._running, "Manticore is already running."
         args = self._args
 
-        replay=None
+        replay = None
         if hasattr(args, 'replay') and args.replay is not None:
             with open(args.replay, 'r') as freplay:
                 replay = map(lambda x: int(x, 16), freplay.readlines())
@@ -705,12 +698,11 @@ class Manticore(object):
             ws_path = None
 
         self._output = ManticoreOutput(ws_path)
-
         self._executor = Executor(initial_state,
-                                  workspace=ws_path,
-                                  policy=self._policy, 
+                                  workspace=self._output.descriptor,
+                                  policy=self._policy,
                                   context=self.context)
-        
+
 
 
         #Link Executor events to default callbacks in manticore object
@@ -723,6 +715,7 @@ class Manticore(object):
         self._executor.subscribe('will_store_state', self._store_state_callback)
         self._executor.subscribe('will_load_state', self._load_state_callback)
         self._executor.subscribe('will_fork_state', self._fork_state_callback)
+        self._executor.subscribe('forking_state', self._forking_state_callback)
         self._executor.subscribe('will_terminate_state', self._terminate_state_callback)
         self._executor.subscribe('will_generate_testcase', self._generate_testcase_callback)
 
@@ -802,4 +795,13 @@ class Manticore(object):
         for cb in self._hooks.get(None, []):
             cb(state)
 
-
+    def _check_disassembler_present(self, disasm):
+        if is_binja_disassembler(disasm):
+            try:
+                import binaryninja
+            except ImportError:
+                err = ("BinaryNinja not found! You MUST own a BinaryNinja version"
+                       " that supports GUI-less processing for this option"
+                       " to work. Please configure your PYTHONPATH appropriately or"
+                       " select a different disassembler")
+                raise SystemExit(err)