+++ CHANGELOG.md
@@ -2,7 +2,34 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
-## [Unreleased](https://github.com/trailofbits/manticore/compare/0.1.3...HEAD)
+## [Unreleased](https://github.com/trailofbits/manticore/compare/0.1.4...HEAD)
+
+## 0.1.4 - 2017-08-18
+
+### Added
+
+- `Manticore.locked_context()` (safe parallel context access)
+- `State.generate_testcase()` (arbitrary testcase generation from hooks)
+- Documentation on [gotchas](http://manticore.readthedocs.io/en/latest/gotchas.html)
+- Command line interface support for symbolic files (`--file`) (thanks [251](https://github.com/251)!)
+- [Experimental] `State.context['branches']` (States track symbolic branches)
+- [Experimental] Support for emulation of [Binary Ninja](https://binary.ninja) IL
+
+### Changed
+
+- Taint parameters added to `State.new_symbolic_buffer()` and `State.symbolicate_buffer()` (thanks [ehennenfent](https://github.com/ehennenfent)!)
+- Improved support for ARM binaries
+- `Manticore.verbosity` logging preset levels
+
+### Fixed
+
+- Numerous bugfixes
+- Fixed workspace error message bug (thanks [chowdaryd](https://github.com/chowdaryd)!)
+- Fixed double workspace bug
+
+### Removed
+
+- [Experimental] `State.generate_inputs()` (superseded by `State.generate_testcase()`)
 
 ## 0.1.3 - 2017-07-14
 
@@ -27,15 +54,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Added
 
-- Function modeling API (`state.invoke_model`, `manticore.variadic`)
+- Function modeling API (`state.invoke_model()`, `manticore.variadic`)
 - `strcmp` and `strlen` models
-- `state.solve_buffer`
+- `state.solve_buffer()`
 - Additional `state` APIs
+- Support for ARMv7 Thumb mode
 
 ### Changed
 
 - Parallel processing API (`m.run(procs)`)
-- `state.solve_n`
+- `state.solve_n()`
 
 ### Fixed